requirements.txt
pandas
numpy
scikit-learn
ipykernel
kaggle
python-dotenv
ipywidgets
python-box
ensure
seaborn
matplotlib
xgboost
mlflow